refactor(core): extract editor subsystem into parish-editor crate
Extract the Parish Designer backend (parish-core/src/editor/, ~1,673 lines
across 8 files: format, live_reload, mod_io, persist, save_inspect, types,
validate, plus the module root) into a new first-class, backend-agnostic
workspace crate `parish-editor`. `parish-core` keeps a thin re-export shim
(`pub use parish_editor as editor;`) so every consumer
(parish-tauri::editor_commands, parish-server::editor_routes,
parish-core::ipc::editor, normalize_mod.rs, input_validation.rs) compiles
with zero import changes.

All editor dependencies were already in leaf crates, so the whole subsystem
moved without a dependency cycle: the game_mod types (AnachronismData,
EncounterTable, FestivalDef, ModManifest, GameMod, world_state_from_mod) live
in parish-mod and WorldState/LocationId in parish-world; live_reload.rs moved
too (imports rewritten crate::game_mod -> parish_mod, crate::world ->
parish_world).

- Register parish-editor in [workspace] members + [workspace.dependencies].
- Add parish-editor to the BACKEND_AGNOSTIC list in architecture_fitness.rs.
- Update crate counts (20 workspace members / 11 leaf crates) and tables in
  AGENTS.md, docs/agent/architecture.md, docs/agent/codebase-map.md, README.md.
- Add the extract-editor verification fixture.

Behavior-preserving: cargo fmt/clippy/test --workspace green; 16 parish-editor
unit tests pass; headless --script transcript structurally identical to the
pre-refactor baseline.
