+++ parish/crates/parish-core/Cargo.toml
@@ -14,6 +14,7 @@ parish-world = { workspace = true }
 parish-palette = { workspace = true }
 parish-npc = { workspace = true }
 parish-mod = { workspace = true }
+parish-editor = { workspace = true }
 parish-persistence = { workspace = true }
 parish-input = { workspace = true }
 tokio = { workspace = true }
