+++ parish/crates/parish-editor/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "parish-editor"
+version = "0.1.0"
+edition = "2024"
+license = "GPL-3.0-only"
+publish = false
+description = "Parish Designer backend: mod browsing, NPC/location editing, cross-reference validation, deterministic persistence, and save-file inspection"
+
+[dependencies]
+parish-types = { workspace = true }
+parish-world = { workspace = true }
+parish-npc = { workspace = true }
+parish-mod = { workspace = true }
+parish-persistence = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+toml = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
