+++ parish/crates/parish-editor/src/lib.rs
@@ -1,12 +1,16 @@
-//! Parish Designer — editor support module.
+//! Parish Designer — editor support crate.
 //!
 //! Backend for the GUI editor utility that lets game designers browse mods,
 //! edit NPC and location data, validate cross-references, and inspect save
 //! files. The editor operates on a **fresh in-memory copy loaded from disk**
-//! and never touches the live gameplay [`GameMod`](crate::game_mod::GameMod)
-//! or [`AppState`]; see `docs/design/designer-editor.md` for the full design.
+//! and never touches the live gameplay [`GameMod`](parish_mod::GameMod) or the
+//! host application state; see `docs/design/designer-editor.md` for the full
+//! design.
 //!
-//! Unlike [`GameMod::load`](crate::game_mod::GameMod::load), which is
+//! Re-exported from `parish-core` as `parish_core::editor` so existing
+//! consumers keep their import paths; the canonical home is this crate.
+//!
+//! Unlike [`GameMod::load`](parish_mod::GameMod::load), which is
 //! all-or-nothing, the editor loads each mod file independently via
 //! [`mod_io::load_mod_snapshot`] so a broken `festivals.json` doesn't hide a
 //! working `npcs.json`. Post-save revalidation uses
