+++ parish/testing/fixtures/play_extract-editor.txt
@@ -1,0 +1,21 @@
+# Verification fixture for extract-editor.
+#
+# This is a behavior-preserving structural refactor (parish-core::editor
+# extracted into a new parish-editor crate). No gameplay behavior changes.
+#
+# The fixture exercises the game engine end-to-end to confirm:
+#   - The mod loads correctly (world graph, NPCs, locations intact)
+#   - Movement, look, status, map, and help still work
+#   - No panics or regressions introduced by the crate restructure
+#
+# Run:
+#   cargo run --manifest-path parish/Cargo.toml -p parish-engine -- \
+#     --headless --script parish/testing/fixtures/play_extract-editor.txt
+#
+# Expected: JSON output for each command with kind matching the command type.
+# Exit code 0. No "Error" or "panic" lines.
+look
+/status
+/map
+/help
+/npcs
